refact(maccel) convert ma-cpi from float--> int >
because it is decimals not really needed.

+++ maccel/maccel.h
@@ -11,7 +11,7 @@ report_mouse_t pointing_device_task_maccel(report_mouse_t mouse_report);
 bool           process_record_maccel(uint16_t keycode, keyrecord_t *record, uint16_t toggle, uint16_t cpi, uint16_t takeoff, uint16_t growth_rate, uint16_t offset, uint16_t limit);
 
 typedef struct _maccel_config_t {
-    float cpi;
+    uint16_t cpi;
     float growth_rate;
     float offset;
     float limit;
@@ -25,12 +25,12 @@ void maccel_enabled(bool enable);
 bool maccel_get_enabled(void);
 void maccel_toggle_enabled(void);
 
-float maccel_get_cpi(void);
+uint16_t maccel_get_cpi(void);
 float maccel_get_takeoff(void);
 float maccel_get_growth_rate(void);
 float maccel_get_offset(void);
 float maccel_get_limit(void);
-void  maccel_set_cpi(float val);
+void  maccel_set_cpi(uint16_t val);
 void  maccel_set_growth_rate(float val);
 void  maccel_set_offset(float val);
 void  maccel_set_limit(float val);
